Fixed-point math added for quad-tree

+++ Runtime/Modules/NetworkModule/NetworkModule.cs
@@ -783,7 +783,7 @@ namespace ME.ECS.Network {
             var tick = this.world.GetCurrentTick();
 
             var timeSinceGameStart = (long)(this.world.GetTimeSinceStart() * 1000L);
-            var targetTick = (Tick)System.Math.Floor(timeSinceGameStart / (this.world.GetTickTime() * 1000d));
+            var targetTick = (Tick)System.Math.Floor(timeSinceGameStart / (double)(this.world.GetTickTime() * 1000d));
             var oldestEventTick = this.statesHistoryModule.GetAndResetOldestTick(tick);
             //UnityEngine.Debug.LogError("Tick: " + tick + ", timeSinceGameStart: " + timeSinceGameStart + ", targetTick: " + targetTick + ", oldestEventTick: " + oldestEventTick);
             if (oldestEventTick == Tick.Invalid || oldestEventTick >= tick) {
